One third refacto...
